// kit/modules/widget/src/data-flow/table/type.ts
export type TableWrapperProps = StyleProps<typeof TableStyle.wrapper> &
  JSX.SharedProps;

/*####################################(TABLE HEADER PROPS)####################################*/

export type TableHeaderProps = StyleProps<typeof TableStyle.header> &
  JSX.SharedProps;

/*####################################(TABLE HEADER PROPS)####################################*/

export type TableHeaderColumnProps = StyleProps<typeof TableStyle.head> &
  JSX.SharedProps;

/*####################################(TABLE HEADER BAR PROPS)####################################*/

export interface TableHeaderBarPaginationProps extends ButtonProps {
  display?: boolean;
  prev?: ButtonProps;
  next?: ButtonProps;
}

export interface TableHeaderBarFilterProps extends ButtonProps {
  display?: boolean;
}

export interface TableHeaderBarSearchProps extends InputFieldProps {
  display?: boolean;
}

export interface TableHeaderBarActionsProps {
  display?: boolean;
  customActions?: ButtonProps[];
  importExport?: ButtonProps | boolean;
  newEntry?: ButtonProps | boolean;
}

export type TableHeaderBarProps = StyleProps<typeof TableStyle.head> &
  JSX.SharedProps & {
    pagination?: TableHeaderBarPaginationProps;
    filter?: TableHeaderBarFilterProps;
    search?: TableHeaderBarSearchProps;
    actions?: TableHeaderBarActionsProps;
  };
// ...
export type TableListProps = StyleProps<typeof TableStyle.list> &
  JSX.SharedProps & {
    each?: ListProps["each"];
    track?: ListProps["track"];
  };

/*####################################(TABLE FOOTER PROPS)####################################*/

export type TableFooterProps = StyleProps<typeof TableStyle.footer> &
  JSX.SharedProps;

/*####################################(TABLE ROW PROPS)####################################*/

export type TableRowProps = StyleProps<typeof TableStyle.row> & JSX.SharedProps;

/*####################################(TABLE CELL PROPS)####################################*/

export type TableCellProps = StyleProps<typeof TableStyle.cell> &
  JSX.SharedProps;

/*####################################(TABLE CAPTION PROPS)####################################*/

export type TableCaptionProps = StyleProps<typeof TableStyle.caption> &
  JSX.SharedProps;